Fixed safe execution, removed boost library, a huge amount of bug fixes and refactoring.

+++ threaddatamanager/getthreadmemorylimit.cpp
@@ -1,5 +1,5 @@
 #include "threaddatamanager/threaddatamanager.h"
 
-double ThreadDataManager::getThreadMemoryLimit(int index){
+long ThreadDataManager::getThreadMemoryLimit(int index){
     return threadMemoryLimits[index];
 }
