+++ threaddatamanager/setthreadmemorylimit.cpp
@@ -1,5 +1,5 @@
 #include "threaddatamanager/threaddatamanager.h"
 
-void ThreadDataManager::setThreadMemoryLimit(int index,double newLimit){
-    threadMemoryLimits[index]=newLimit;
+void ThreadDataManager::setThreadMemoryLimit(int index, long newLimit){
+    threadMemoryLimits[index] = newLimit;
 }
