+++ threaddatamanager/setthreadtimelimit.cpp
@@ -1,8 +1,6 @@
-#include <chrono>
-
 #include "threaddatamanager/threaddatamanager.h"
 
 void ThreadDataManager::setThreadTimeLimit(int index,
-    int newLimit){
-    threadTimeLimits[index]=newLimit;
+    long newLimit){
+    threadTimeLimits[index] = newLimit;
 }
